Revert "refactor: inject Slack session launcher dependencies"
This reverts commit 6d2d8cd710b971cca9e9f42c0cf3b6cec456b57e.
